refactor: reorganize props and locations of markdown renderers
