--- frontend/src/components/markdown-renderer/common-markdown-renderer-props.ts
@@ -1,15 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2022 The HedgeDoc developers (see AUTHORS file)
- *
- * SPDX-License-Identifier: AGPL-3.0-only
- */
-import type { Ref } from 'react'
-
-export interface CommonMarkdownRendererProps {
-  baseUrl: string
-  outerContainerRef?: Ref<HTMLDivElement>
-  newlinesAreBreaks?: boolean
-  lineOffset?: number
-  className?: string
-  markdownContentLines: string[]
-}
